rename firestore proxies
