+++ src/main/java/fr/trendev/comptandye/security/controllers/jwt/dto/firestore/FirestoreJWTRevokedSetProxy.java
@@ -5,15 +5,14 @@
  */
 package fr.trendev.comptandye.security.controllers.jwt.dto.firestore;
 
-import fr.trendev.comptandye.security.entities.JWTWhiteMapEntry;
+import fr.trendev.comptandye.security.entities.JWTRecord;
 import java.io.Serializable;
-import java.util.List;
+import java.util.Set;
 import java.util.concurrent.CompletionStage;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -24,29 +23,27 @@ import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
  *
  * @author jsie
  */
-@Path("jwtwhitemap")
+@Path("jwtrevokedset")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @RegisterRestClient
-public interface FirestoreJWTWhiteMapProxyService extends Serializable {
+public interface FirestoreJWTRevokedSetProxy extends Serializable {
 
     @GET
-    CompletionStage<List<JWTWhiteMapEntry>> getAll() throws
+    CompletionStage<Set<JWTRecord>> getAll() throws FirestoreProxyException;
+
+    @POST
+    CompletionStage<JWTRecord> create(JWTRecord record) throws
             FirestoreProxyException;
 
     @POST
-    CompletionStage<JWTWhiteMapEntry> create(JWTWhiteMapEntry jwtWhiteMapEntry)
-            throws
-            FirestoreProxyException;
-
-    @PUT
-    CompletionStage<JWTWhiteMapEntry> update(JWTWhiteMapEntry jwtWhiteMapEntry)
-            throws
+    @Path("bulk-creation")
+    CompletionStage<Set<JWTRecord>> bulkCreation(Set<JWTRecord> records) throws
             FirestoreProxyException;
 
     @DELETE
-    @Path("{email}")
-    CompletionStage<String> delete(@PathParam("email") String email) throws
+    @Path("{token}")
+    CompletionStage<String> delete(@PathParam("token") String token) throws
             FirestoreProxyException;
 
 }
