+++ src/main/java/fr/trendev/comptandye/security/controllers/jwt/dto/firestore/FirestoreJWTWhiteMapProxy.java
@@ -5,14 +5,15 @@
  */
 package fr.trendev.comptandye.security.controllers.jwt.dto.firestore;
 
-import fr.trendev.comptandye.security.entities.JWTRecord;
+import fr.trendev.comptandye.security.entities.JWTWhiteMapEntry;
 import java.io.Serializable;
-import java.util.Set;
+import java.util.List;
 import java.util.concurrent.CompletionStage;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -23,27 +24,29 @@ import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
  *
  * @author jsie
  */
-@Path("jwtrevokedset")
+@Path("jwtwhitemap")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @RegisterRestClient
-public interface FirestoreJWTRevokedSetProxyService extends Serializable {
+public interface FirestoreJWTWhiteMapProxy extends Serializable {
 
     @GET
-    CompletionStage<Set<JWTRecord>> getAll() throws FirestoreProxyException;
-
-    @POST
-    CompletionStage<JWTRecord> create(JWTRecord record) throws
+    CompletionStage<List<JWTWhiteMapEntry>> getAll() throws
             FirestoreProxyException;
 
     @POST
-    @Path("bulk-creation")
-    CompletionStage<Set<JWTRecord>> bulkCreation(Set<JWTRecord> records) throws
+    CompletionStage<JWTWhiteMapEntry> create(JWTWhiteMapEntry jwtWhiteMapEntry)
+            throws
+            FirestoreProxyException;
+
+    @PUT
+    CompletionStage<JWTWhiteMapEntry> update(JWTWhiteMapEntry jwtWhiteMapEntry)
+            throws
             FirestoreProxyException;
 
     @DELETE
-    @Path("{token}")
-    CompletionStage<String> delete(@PathParam("token") String token) throws
+    @Path("{email}")
+    CompletionStage<String> delete(@PathParam("email") String email) throws
             FirestoreProxyException;
 
 }
